requirements.txt
pyyaml>=6.0
PyGithub>=2.1
python-dotenv>=1.0.0
gitpython>=3.1.40
requests>=2.31.0
questionary>=2.0
rich>=13.0
